64-bit fixes

+++ src/pd/red-black/main.c
@@ -23,8 +23,8 @@ int icomp(char *i, char *j)
 {
   int I, J;
   
-  I = (int) i;
-  J = (int) j;
+  I = (int) (long) i;
+  J = (int) (long) j;
   if (I == J) return 0;
   if (I > J) return -1; else return 1;
 }
@@ -49,12 +49,12 @@ main(int argc, char **argv)
     if (i > 0) {
       j = random()%i;
       
-      rb_delete_node(rb_find_gkey(argt, (char *) (a[j]), icomp));
+      rb_delete_node(rb_find_gkey(argt, (char *) (long) (a[j]), icomp));
       a[j] = random() % 1000;
-      rb_insertg(argt, (char *) (a[j]), NULL, icomp);
+      rb_insertg(argt, (char *) (long) (a[j]), NULL, icomp);
     }
     a[i] = random() % 1000;
-    rb_insertg(argt, (char *) (a[i]), NULL, icomp);
+    rb_insertg(argt, (char *) (long) (a[i]), NULL, icomp);
   }
   tb = 0;
   mxp = 0;
